♻️ Refactor code
detach App component from index.jsx
