+++ src/index.jsx
@@ -1,34 +1,8 @@
-import React, { useState } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
+import App from './App';
 
-import Buttons from './Buttons';
-import ClickMeCounter from './ClickMeCounter';
-
-function App() {
-  const [state, setState] = useState({
-    count: 0,
-  });
-
-  const { count } = state;
-
-  function handleClickButton(i) {
-    setState({
-      count: count + i,
-    });
-  }
-
-  return (
-    <>
-      <ClickMeCounter
-        count={count}
-        onClick={handleClickButton}
-      />
-
-      <Buttons
-        onClick={handleClickButton}
-      />
-    </>
-  );
-}
-
-ReactDOM.render(<App />, document.getElementById('app'));
+ReactDOM.render(
+  <App />,
+  document.getElementById('app'),
+);
